+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Next
+
+- Added backlight compensation `camera.backlightCompensation`.
+- Added powerline frequency `camera.powerlineFrequency`.
+- Can now set focus and focus auto, not just get the values.
+
 ## 1.0.0
 
 - Initial version with support for pan, tilt, zoom, and focus. Runs on macOS and Windows.